server/requirements.txt
alembic==1.12.1
blinker==1.6.3
cffi==1.16.0
click==8.1.7
cmdstanpy==1.2.0
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.43.1
greenlet==3.0.1
holidays==0.35
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
prophet==1.1.5
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
stanio==0.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
